Delete src/lib: move remaining helpers to their consumers
After this collapse, src/lib/ goes away entirely. Each helper now lives
in the single src/ or test/ file that uses it:

- `eigendecomp` and `map_eigvals` move from `lib/ITensorsExtensions/`
  into `src/apply.jl` (only consumer in `src/` was `simple_update_bp`).
  Both are now module-level `ITensorNetworks` functions; `src/apply.jl`
  drops the `ITensorsExtensions.` qualifier on its four call sites.

- `tensor_sum` moves from `lib/ITensorsExtensions/` into `src/opsum.jl`
  (only consumer was `ITensors.ITensor(::Sum, ::IndsNetwork)` in the
  same file).

- The `indtype(link_space, site_space)` 4-form, `indtype(space)`,
  `_indtype` family, and `IsIndexSpace` move from
  `lib/ITensorsExtensions/` into `src/indsnetwork.jl` (only consumer
  was the `IndsNetwork{V, I}(g, link_space, site_space)` constructor).

- `maybe_real` moves from `lib/BaseExtensions/` into `src/opsum.jl`
  (consumers were `src/opsum.jl` and `src/apply.jl`; both files reach
  it as a top-level `ITensorNetworks` function now). `to_tuple` was
  only used in `test/utils.jl::group_terms` and moves there directly.

- `test/test_itensorsextensions.jl` renamed to `test/test_map_eigvals.jl`,
  imports updated from `ITensorNetworks.ITensorsExtensions: ...` to
  `ITensorNetworks: ...`. The testset name changes from
  `"ITensorsExtensions"` to `"map_eigvals"`.

- `include("lib/BaseExtensions/...")` removed from
  `src/ITensorNetworks.jl`; `src/lib/` directory deleted outright.

Net effect: drops two submodules' worth of indirection without losing
any in-tree behavior. All tests pass.
