+++ src/ITensorNetworks.jl
@@ -1,7 +1,5 @@
 module ITensorNetworks
 
-include("lib/BaseExtensions/src/BaseExtensions.jl")
-include("lib/ITensorsExtensions/src/ITensorsExtensions.jl")
 include("visualize.jl")
 include("abstractindsnetwork.jl")
 include("indextags.jl")
