--- src/lib/ITensorsExtensions/src/ITensorsExtensions.jl
@@ -1,4 +1,0 @@
-module ITensorsExtensions
-include("itensor.jl")
-include("itensor_more.jl")
-end
